Fix: Separated all logic by ui
