+++ src/components/Sidebar/index.jsx
@@ -1,42 +1,13 @@
 "use client"
 import React from 'react'
-import Cookies from 'js-cookie'
 import Button from '@/components/Button'
-import { useRouter } from 'next/navigation'
 import UserInfo from '@/components/UserInfo'
-import { useItem } from '@/context'
+import useSidebar from '@/hooks/useSidebar'
 import SideHeader from '@/components/SideHeader'
 
 const Index = () => {
 
-    const router = useRouter()
-    const styles = ['bg-secondary hover:bg-auxiliar hover:text-secondary', 'bg-white hover:bg-auxiliar !text-black']
-    const [user, setUser] = React.useState()
-    const { item, setItem } = useItem() 
-
-    React.useEffect(() => {
-      try {
-        const userLogged = JSON.parse(Cookies?.get('SessionInfo'))
-
-        if (userLogged) {
-          setUser(userLogged?.answer[0])
-        }
-      } catch (error) {
-        console.error(error)
-      }
-    }, [])
-
-    const handleLogout = () => {
-      Cookies.remove('SessionInfo')
-      setTimeout(() => {
-        router.push('/')
-      }, 2000);
-    }
-
-    const handleChange = (id) => {
-      router.push('/main')
-      setItem(id)
-    }
+    const { styles, user, item, handleChange, handleLogout } = useSidebar()
 
   return (
     <aside className="bg-primary w-1/6 h-screen relative py-12 rounded-r-[4rem]">
