+++ src/containers/ComercialInfo/index.jsx
@@ -3,7 +3,7 @@ import React from 'react'
 import Image from 'next/image'
 import Cookies from 'js-cookie'
 import { useRouter } from 'next/navigation'
-import { formatPrice } from '@/utils/formatPrice'
+import { formatPrice } from '@/utils'
 import CardSection from '@/components/CardSection'
 import LinkSection from '@/components/LinkSection'
 import './index.css'
