+++ src/containers/Login/index.jsx
@@ -1,53 +1,15 @@
 "use client"
-import axios from 'axios';
 import React from 'react';
+import useLogin from '@/hooks/useLogin';
+import Loader from '@/components/Loader';
 import Button from '@/components/Button';
+import SideHeader from '@/components/SideHeader';
 import LoginSection from '@/components/LoginSection'
 import PasswordSection from '@/components/PasswordSection';
-import { useRouter } from 'next/navigation';
-import Loader from '@/components/Loader';
-import Cookies from 'js-cookie';
-import SideHeader from '@/components/SideHeader';
-
 
 const Index = () => {
 
-    const router = useRouter()
-    const [loaderActive, setLoaderActive] = React.useState(false)
-    const [alert, setAlert] = React.useState('')
-
-    const [formData, setFormData] = React.useState({
-        Correo: '',
-        Contraseña: '',
-    }); 
-
-    const handleInputChange = (e) => {
-        const { id, value } = e.target;
-        setFormData({
-        ...formData,
-        [id]: value,
-        });
-    };
-
-    const eventLogin = (response) => {
-        router.push('/main')
-        Cookies.set('SessionInfo', JSON.stringify(response.data))
-        setLoaderActive(false)
-    }
-
-    const eventLoginFailed = (error) => {
-        setLoaderActive(false)
-        setAlert(error?.response?.data)
-    }
-
-    const onLoginSubmit = (e) => {
-        setAlert('')
-        e.preventDefault()
-        setLoaderActive(true)
-        axios.post(`${process.env.BACK_LINK}/api/loginUser`, formData)
-        .then((response) => eventLogin(response))
-        .catch((error) => eventLoginFailed(error))
-    }
+    const { loaderActive, alert, handleInputChange, onLoginSubmit, formData } = useLogin()
 
   return (
     <form className="flex flex-col gap-4 bg-auxiliar p-6 rounded-lg" onSubmit={onLoginSubmit}>
