+++ src/containers/ResidencialInfo/index.jsx
@@ -4,7 +4,7 @@ import axios from 'axios'
 import Image from 'next/image'
 import Cookies from 'js-cookie'
 import { useRouter } from 'next/navigation'
-import { formatPrice } from '@/utils/formatPrice'
+import { formatPrice } from '@/utils'
 import LinkSection from '@/components/LinkSection'
 import CardSection from '@/components/CardSection'
 import './index.css'
