+++ src/hooks/useObservation.js
@@ -1,0 +1,53 @@
+import React from 'react'
+import axios from 'axios'
+import Cookies from 'js-cookie'
+
+const useObservation = (setState, id, letter, type) => {
+  
+    const availableStatus = ['Pendiente', 'Atendido', 'Descartado']
+    const [text, setText] = React.useState(null)
+    const [status, setStatus] = React.useState('Pendiente')
+    const [loaderActive, setLoaderActive] = React.useState(false)
+
+    const handleChange = (e) => {
+        setText(e.target.value)
+    }
+
+    const handleChangeList = (e) => {
+        setStatus(e.target.value)
+    }
+
+    const handleSubmit = (e) => {
+        e.preventDefault()
+        setLoaderActive(true)
+        const sessionInfo = JSON.parse(Cookies?.get('SessionInfo'))
+        try {
+            Promise.all([
+                axios.put(
+                `${process.env.BACK_LINK}/api/observacion${letter}/${id}`, 
+                { observacion: text }, {
+                    headers: {
+                        "Authorization": `Bearer ${sessionInfo?.token}`
+                    }
+                }),
+                axios.put(
+                `${process.env.BACK_LINK}/api/${type}/updateRevisado/${id}`, 
+                { newStatus: availableStatus.indexOf(status) }, {
+                    headers: {
+                        "Authorization": `Bearer ${sessionInfo?.token}`
+                    }
+                })
+            ])
+            .then(() => setLoaderActive(false), location.reload())
+            .catch(() => setLoaderActive(false), location.reload())
+            setState(false)
+        } catch(error) {
+            console.error(error)
+            setState(false)
+        }
+    }
+
+    return { availableStatus, text, status, loaderActive, handleChange, handleChangeList, handleSubmit }
+}
+
+export default useObservation
