--- src/utils/formatPrice.js
@@ -1,7 +1,0 @@
-export const formatPrice = (number) => {
-    return number?.toLocaleString('es-CO', {
-        style: 'currency',
-        currency: 'COP',
-        minimumFractionDigits: 0
-      });
-};
