--- src/utils/getDate.js
@@ -1,4 +1,0 @@
-export const getDate = () => {
-    const date = new Date()
-    return date.getDate()
-}
